Moving some of redhat/infra_manager code to seperate model
This is just a small step in the direction of making this
more object oriented, for now are least this cosmetic change
is done to help work on another PR.
